CARS007/trabalho-1-bibliometria/requirements.txt
python-dotenv==1.1.1
pandas==2.3.2
matplotlib==3.10.5
geopy==2.4.1
plotly==6.3.0
nbformat>=4.2.0